alg.cpp_upd3

+++ src/alg.cpp
@@ -27,7 +27,7 @@ unsigned int collatzLen(uint64_t num) {
     if (num == 0) return 0;
     if (num == 1) return 1;
 
-    unsigned int length = 1;
+    uint64_t length = 1;
     uint64_t currentN = num;
 
     while (currentN != 1) {
@@ -42,7 +42,7 @@ unsigned int collatzLen(uint64_t num) {
     return length;
 }
 
-unsigned int seqCollatz(unsigned int *maxlen,
+unsigned int seqCollatz(uint64_t *maxlen,
                         uint64_t lbound,
                         uint64_t rbound) {
     if (lbound > rbound || lbound < 2) {
@@ -50,8 +50,8 @@ unsigned int seqCollatz(unsigned int *maxlen,
         return 0;
     }
 
-    unsigned int topLength = 0;
-    unsigned int bestNumber = lbound;
+    uint64_t topLength = 0;
+    uint64_t bestNumber = lbound;
 
     for (uint64_t num = lbound; num <= rbound; ++num) {
         uint64_t len = collatzLen(num);
